AWS-CustomerServ/backend/CustomerService/lambda/chat/requirements.txt
python-dotenv==1.0.0
boto3==1.34.69
